fix(providers): keep display helpers tool agnostic
